# backend_go/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login, create new session, return auth cookies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login in system",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session token",
                        "schema": {
                            "$ref": "#/definitions/schemas.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "description": "Getting notify categories which user created",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Getting notify categories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Session",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "schema": {
                            "$ref": "#/definitions/schemas.EmberCategories"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifies": {
            "get": {
                "description": "Getting notify  which user created",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Getting notifies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id for select only linked notifies",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "Session",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of notifies",
                        "schema": {
                            "$ref": "#/definitions/schemas.EmberNotifies"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api_utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api_utils.ErrorData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "api_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api_utils.ErrorData"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Notify": {
            "type": "object",
            "properties": {
                "acceptors": {},
                "autodisable_at": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_autodisable": {
                    "type": "boolean"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "is_disabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "periodic": {
                    "type": "string"
                },
                "repeate_interval": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "schemas.EmberCategories": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Category"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/schemas.EmberMeta"
                }
            }
        },
        "schemas.EmberMeta": {
            "type": "object",
            "properties": {
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "schemas.EmberNotifies": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/schemas.EmberMeta"
                },
                "notify": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Notify"
                    }
                }
            }
        },
        "schemas.LoginParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "schemas.SessionResponse": {
            "type": "object",
            "properties": {
                "session": {
                    "type": "string"
                }
            }
        }
    }
}